ports/psoc6: Removed psoc6_gpio driver.
