+++ ports/psoc6/Makefile
@@ -128,7 +128,6 @@ DRIVERS_SRC_C += $(addprefix drivers/,\
 	)
 
 DRIVERS_SRC_C += $(addprefix drivers/,\
-	machine/psoc6_gpio.c \
 	machine/psoc6_i2c.c \
 	machine/psoc6_pwm.c \
 	machine/psoc6_system.c  \
