+++ ports/psoc6/modules/machine/machine_spi.c
@@ -13,7 +13,6 @@
 
 
 // port-specific includes
-#include "drivers/machine/psoc6_gpio.h"
 #include "modmachine.h"
 #include "mplogger.h"
 #include "pins.h"
